Refactoring geography and motion types.
