+++ src/upstage/geography/__init__.py
@@ -4,8 +4,13 @@
 # See the LICENSE file in the project root for complete license terms and disclaimers.
 """Geographical methods for intersections, distances, and locations."""
 
-from .intersections import (
+from .geo_types import (
+    INTERSECTION_LOCATION_CALLABLE,
+    LAT_LON,
     LAT_LON_ALT,
+    EarthProtocol,
+)
+from .intersections import (
     CrossingCondition,
     get_intersection_locations,
 )
@@ -13,9 +18,12 @@ from .spherical import Spherical
 from .wgs84 import WGS84
 
 __all__ = [
+    "EarthProtocol",
     "Spherical",
     "WGS84",
     "get_intersection_locations",
+    "LAT_LON",
     "LAT_LON_ALT",
     "CrossingCondition",
+    "INTERSECTION_LOCATION_CALLABLE",
 ]
